+++ deid/dicom/pixels/detect.py
@@ -2,7 +2,7 @@
 
 detect.py: functions for pixel scrubbing
 
-Copyright (c) 2017-2021 Vanessa Sochat
+Copyright (c) 2017-2022 Vanessa Sochat
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal