+++ deid/tests/test_clean_pixel_dimensions.py
@@ -3,7 +3,7 @@
 """
 Test DICOM Cleaner - Images with varying pixel dimensions
 
-Copyright (c) 2016-2021 Vanessa Sochat
+Copyright (c) 2016-2022 Vanessa Sochat
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal