+++ deid/tests/test_dicom_groups.py
@@ -3,7 +3,7 @@
 """
 Testing groups for a deid recipe (values and fields)
 
-Copyright (c) 2020-2021 Vanessa Sochat
+Copyright (c) 2020-2022 Vanessa Sochat
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal